// data/types/books.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Books = {
  bookId: string;
  chapters: string[];
  englishDivision:
    | "pentateuch"
    | "historical"
    | "wisdom"
    | "majorProphets"
    | "minorProphets"
    | "gospels"
    | "acts"
    | "epistles"
    | "revelation";
  hebrewDivision: "torah" | "nevi'im" | "ketuvim" | null;
  hebrewOrder: number | null;
  name: string;
  order: number;
  testament: "old" | "new";
}[];